TM D2S7 rebuilt
git-svn-id: https://sfs.atcorau.local/svn/SoftwareRepository/trunk/PN022@491 a8a9b14d-6b92-5b4b-9153-4cd90bb22c97

+++ Scor/dal/dal/DalDataBuffer.h
@@ -29,7 +29,7 @@ namespace AtCor{
 							unsigned short tonometerData; /**< Tonometer data */   
 							unsigned short cuffPulseData; /**< Cuff Pulse data */
 							unsigned short cuffPressure; /**< Cuff Pressure data */
-							unsigned short countdownTimer; /**< Countdown Timer interval */
+							unsigned long countdownTimer; /**< Countdown Timer interval */
 
 							/**
 							* Constructor for the structure. @n
